requirements.txt
streamlit
numpy
pandas
matplotlib
scipy
numba
Pillow
